requirements.txt
blinker==1.9.0
click==8.3.0
exceptiongroup==1.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.1
SQLAlchemy==2.0.44
tomli==2.3.0
typing_extensions==4.15.0
Werkzeug==3.1.3
blinker==1.9.0
click==8.3.0
exceptiongroup==1.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.1
SQLAlchemy==2.0.44
tomli==2.3.0
typing_extensions==4.15.0
Werkzeug==3.1.3
